+++ CHANGELOG.md
@@ -36,6 +36,7 @@
   - pkg: new `Config` interface type to support new generic `FromViper` config file parsing
 - changed flags `--k3s-server-arg` & `--k3s-agent-arg` into `--k3s-arg` with nodefilter support (#605)
   - new config path `options.k3s.extraArgs`
+- config file: environment variables (`$VAR`, `${VAR}` will be expanded unconditionally) (#643)
 
 ### Misc